refactor(logger_core): logger_core is now a 'tools'
